+++ app/assets/javascripts/discourse/tests/acceptance/topic-footer-dropdown-api-test.js
@@ -3,11 +3,16 @@ import { visit } from "@ember/test-helpers";
 import { test } from "qunit";
 import { withPluginApi } from "discourse/lib/plugin-api";
 import selectKit from "discourse/tests/helpers/select-kit-helper";
+import { clearTopicFooterDropdowns } from "discourse/lib/register-topic-footer-dropdown";
 
 acceptance(
   "Topic - Plugin API - registerTopicFooterDropdown",
   function (needs) {
     needs.user();
+
+    needs.hooks.afterEach(() => {
+      clearTopicFooterDropdowns();
+    });
 
     test("adds topic footer dropdown through API", async function (assert) {
       const done = assert.async();
